+++ src/frontend/js/widgets/Dashboard/components/DashboardItem/Order/DashboardItemOrder.tsx
@@ -114,8 +114,8 @@ export const DashboardItemOrder = ({
               </div>
               {showDetailsButton && (
                 <RouterButton
+                  size="small"
                   className="dashboard-item__button"
-                  color="transparent-darkest"
                   href={getRoutePath(LearnerDashboardPaths.ORDER, { orderId: order.id })}
                   data-testid="dashboard-item-order__button"
                 >